assignment-01-DA24M011-Nandhakishore/requirements.txt
requests 
beautifulsoup4 
PyYAML 
traceback2 
psycopg2